Add benchmark.
