+++ paddleaudio/tests/benchmark/README.md
@@ -7,7 +7,7 @@ pip install pytest-benchmark
 # 2. Run
 Run the specific script for profiling.
 ```sh
-pytest features.py
+pytest melspectrogram.py
 ```
 
 Result:
@@ -17,31 +17,23 @@ platform linux -- Python 3.7.7, pytest-7.0.1, pluggy-1.0.0
 benchmark: 3.4.1 (defaults: timer=time.perf_counter disable_gc=False min_rounds=5 min_time=0.000005 max_time=1.0 calibration_precision=10 warmup=False warmup_iterations=100000)
 rootdir: /ssd3/chenxiaojie06/PaddleSpeech/DeepSpeech/paddleaudio
 plugins: typeguard-2.12.1, benchmark-3.4.1, anyio-3.5.0
-collected 12 items
+collected 4 items
 
-features.py ............                                                                                                                                          [100%]
+melspectrogram.py ....                                                                                                                                            [100%]
 
 
----------------------------------------------------------------------------------------------------- benchmark: 12 tests ----------------------------------------------------------------------------------------------------
-Name (time in us)                            Min                    Max                   Mean                StdDev                 Median                 IQR            Outliers         OPS            Rounds  Iterations
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-test_melspect_gpu_torchaudio            210.7229 (1.0)         338.5879 (1.0)         217.4949 (1.0)         11.3591 (1.02)        214.0319 (1.0)        8.3707 (1.0)           6;5  4,597.8093 (1.0)         186           1
-test_log_melspect_gpu_torchaudio        375.4422 (1.78)      1,024.8050 (3.03)        387.3589 (1.78)        18.7080 (1.69)        385.2872 (1.80)       9.4259 (1.13)        31;31  2,581.5853 (0.56)       1420           1
-test_mfcc_gpu_torchaudio                422.4107 (2.00)        700.7364 (2.07)        454.9903 (2.09)        47.3926 (4.27)        436.6031 (2.04)      15.4376 (1.84)      159;193  2,197.8493 (0.48)       1078           1
-test_melspect_gpu                       819.3776 (3.89)      1,161.9311 (3.43)        900.9168 (4.14)       147.0245 (13.26)       830.7453 (3.88)     115.4500 (13.79)         1;1  1,109.9805 (0.24)          5           1
-test_log_melspect_gpu                 1,197.9323 (5.68)      1,280.0004 (3.78)      1,214.0182 (5.58)        11.0918 (1.0)       1,211.6358 (5.66)      10.0820 (1.20)        84;31    823.7109 (0.18)        533           1
-test_mfcc_gpu                         1,337.0719 (6.35)      1,601.5675 (4.73)      1,355.4527 (6.23)        26.4458 (2.38)      1,348.6911 (6.30)      13.1410 (1.57)        16;17    737.7609 (0.16)        193           1
-test_melspect_cpu_torchaudio          1,374.8817 (6.52)      3,937.5033 (11.63)     1,574.8930 (7.24)       355.4223 (32.04)     1,409.1432 (6.58)     193.7435 (23.15)       36;49    634.9638 (0.14)        291           1
-test_log_melspect_cpu_torchaudio      1,390.2634 (6.60)      2,121.2976 (6.27)      1,559.3045 (7.17)       220.3090 (19.86)     1,409.4356 (6.59)     349.1524 (41.71)       106;0    641.3116 (0.14)        445           1
-test_mfcc_cpu_torchaudio              1,445.6678 (6.86)      3,801.8432 (11.23)     1,680.8559 (7.73)       395.5443 (35.66)     1,469.8748 (6.87)     305.6149 (36.51)       38;35    594.9350 (0.13)        469           1
-test_melspect_cpu                    20,620.2641 (97.85)    20,984.0760 (61.98)    20,721.4942 (95.27)       70.2757 (6.34)     20,717.8025 (96.80)     57.8668 (6.91)          6;2     48.2591 (0.01)         30           1
-test_log_melspect_cpu                21,025.3932 (99.78)    48,894.0198 (144.41)   23,057.7049 (106.01)   5,440.3207 (490.48)   21,190.5045 (99.01)    190.0699 (22.71)         4;9     43.3695 (0.01)         44           1
-test_mfcc_cpu                        21,127.2798 (100.26)   45,811.5358 (135.30)   23,176.4022 (106.56)   5,041.0751 (454.49)   21,319.1714 (99.61)    149.0396 (17.80)         5;9     43.1473 (0.01)         44           1
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+-------------------------------------------------------------------------------------------------- benchmark: 4 tests -------------------------------------------------------------------------------------------------
+Name (time in us)                        Min                    Max                   Mean              StdDev                 Median                 IQR            Outliers         OPS            Rounds  Iterations
+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+test_melspect_gpu_torchaudio        202.0765 (1.0)         360.6230 (1.0)         218.1168 (1.0)       16.3022 (1.0)         214.2871 (1.0)       21.8451 (1.0)          40;3  4,584.7001 (1.0)         286           1
+test_melspect_gpu                   657.8509 (3.26)        908.0470 (2.52)        724.2545 (3.32)     106.5771 (6.54)        669.9096 (3.13)     113.4719 (5.19)          1;0  1,380.7300 (0.30)          5           1
+test_melspect_cpu_torchaudio      1,247.6053 (6.17)      2,892.5799 (8.02)      1,443.2853 (6.62)     345.3732 (21.19)     1,262.7263 (5.89)     221.6385 (10.15)       56;53    692.8637 (0.15)        399           1
+test_melspect_cpu                20,326.2549 (100.59)   20,607.8682 (57.15)    20,473.4125 (93.86)     63.8654 (3.92)     20,467.0429 (95.51)     68.4294 (3.13)          8;1     48.8438 (0.01)         29           1
+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 
 Legend:
   Outliers: 1 Standard Deviation from Mean; 1.5 IQR (InterQuartile Range) from 1st Quartile and 3rd Quartile.
-    OPS: Operations Per Second, computed as 1 / Mean
-    ========================================================================== 12 passed in 26.81s ==========================================================================
+  OPS: Operations Per Second, computed as 1 / Mean
+========================================================================== 4 passed in 21.12s ===========================================================================
 
 ```
